refactor(page-layout): rename scss files to use .module.scss
