+++ draft-packages/page-layout/KaizenDraft/Skirt/Skirt.module.scss
@@ -1,4 +1,7 @@
-@import "../decision-tokens.scss";
+@import "~@kaizen/design-tokens/sass/color";
+
+$dt-color-background-color-default: $color-purple-600;
+$dt-color-background-color-education: $color-blue-100;
 
 .container {
   position: relative;
