+++ draft-packages/page-layout/KaizenDraft/Skirt/SkirtCard.module.scss
@@ -1,5 +1,5 @@
 @import "~@kaizen/design-tokens/sass/layout";
-@import "~@kaizen/draft-page-layout/KaizenDraft/Content/styles.scss";
+@import "../variables";
 
 .wrapper {
   @media (max-width: $layout-breakpoints-large) {
