+++ draft-packages/page-layout/KaizenDraft/_variables.scss
@@ -1,0 +1,3 @@
+// This replicates the layout max-width logic in
+// draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/TitleBlockZen.scss
+$content-margin-width-on-medium-and-small: 12px;
