--- draft-packages/page-layout/KaizenDraft/decision-tokens.scss
@@ -1,3 +1,0 @@
-@import "~@kaizen/design-tokens/sass/color";
-$dt-color-background-color-default: $color-purple-600;
-$dt-color-background-color-education: $color-blue-100;
